requirements.txt
pyTelegramBotAPI==4.8.0
websocket-client==1.4.2
pycoingecko==3.1.0